requirements.txt
# Core
anthropic>=0.42.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Google APIs
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0

# Vector Store (for social media voice matching)
chromadb>=0.4.0

# Database
# sqlite3 is built-in

# PDF Generation (for invoices)
reportlab>=4.0.0

# UI
streamlit>=1.30.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
